requirements.txt
falcon==3.0.0
pandas==1.3.3
numpy==1.21.2
boto3==1.18.33
pandas==1.3.3
pymysql==1.0.2
mysql-connector-python==8.0.28
